+++ src/api/index.ts
@@ -1,6 +1,16 @@
-import { ChatMessage } from '../types';
-
 export interface APIClient {
   fetchChat(messages: ChatMessage[]): AsyncGenerator<string | undefined>;
   fetchCompletions(prefix: string, suffix: string): Promise<string | undefined>;
+}
+
+export type ChatRole = 'system' | 'assistant' | 'user';
+
+export interface ChatMessage {
+  role: ChatRole;
+  content: string;
+}
+
+export interface ChatHistory {
+  messages: ChatMessage[];
+  response: string;
 }
