+++ src/api/proxies/memory-cache.ts
@@ -1,7 +1,6 @@
 import { createHash } from 'crypto';
 import Markpilot from 'src/main';
-import { APIClient } from '..';
-import { ChatMessage } from '../../types';
+import { APIClient, ChatMessage } from '..';
 
 export class MemoryCacheProxy implements APIClient {
   private store: Map<string, string> = new Map();
