+++ src/settings/migrators/index.ts
@@ -1,0 +1,1 @@
+export type SettingsMigrator<From, To> = (settings: From) => To;
